requirements.txt
slack-bolt>=1.18.0
openai>=1.6.0
python-dotenv>=1.0.0